add "_" Underscore data type
This allows us to describe computations without evaluating them.

I initially developed this to make writing schema (pre) conditions more concise & readable, but then I realized this is a general form of describing computation.

In other words, its essentially a new way of (limited) writing functions.

This is very nice IMO, and it can be extended very easily.

One thing that we could do with this new syntax is do optimizations.

For example

_::map(:name)::map(to_lowercase)

if we somehow have it known that consecutive "map" calls could be made into one via function composition we can drastically improve performance

In other words, I believe this could become the foundation for transducers in protej.

One way we could do that is say

define Composable for map = true

and we could do something in the ::call of Underscore to see if we have consecutive "composable" functions, then just compose them.

This would require a slight change to how Underscore protocols are implemented. Currently its just a series of functions that will be composed, but we could do the following

new Underscore([{func: map, args: [f]])

for example for map, and likewise for everything, store the function & arguments separately. This way when they need to be optimized they can be pulled apart.