Exec (array comprehension)
The Vector syntax is proving to be a great primitive for creating powerful DSLs

Along with _, Exec feels just as powerful as python for comprehensions, with easy to extend the syntax.

See the :verify for a novel idea.

Its all very concise & works so well because its just data as rich hickey would say.

I really see the power in separating data from implementation now.