The clojure based compiler is quite slow now though & I thought it might be JVM startup but it seems to be tied to the length of the source input.. which isn't great.
With the prelude a compile takes 15seconds, and just the 6 line test file is 5seconds.

Having the test file take 6 seconds is awful enough, so I thought it might be jvm. but damn, no its slow.

I'm not a clojure expert so I wouldn't be surprised if there's some low hanging fruit without fundamentally rewriting my compiler.

I'm not familiar with the performance tradeoffs in clojure nearly as well I am with JavaScript.

I don't believe I'm doing things that would provide such an insane difference, but in my javascript compiler, a small file compiles in 0.23 seconds!! >:( AND to make it even worse, the parser (827 lines) takes 0.48s to compile!!! Jesus christ.

This as well as having a stable js-runnable compiler is what made me consider rewriting coil in coil.

Compile times really shouldn't be an issue with the source I'm dealing with, but god damn it is, even 5 seconds make me want to bash my head sometimes.

clojurescript isn't promising on performance either & the build steps also make me want to bash my head into a wall.

I love clojure, and the parser dsl I wrote in it is by far the most stable of any language I've used.

I'm not sure how much faster it'll be in coil (if at all), but it can't be this god damn slow.

Also for rewriting in coil, I feel much more confident finding out performance bottlenecks & dismantling them.

I hope this isn't good-bye to clojure, but it might be a break <3

Rewriting in coil, might also prove to be a huge pain, but it might be time to start giving it a chance.