I did some quick non scientific tests with the new lexer & parser. Assuming most of the time in clojure is spent in the parser & not in codegen.. the results are looking very promising
500 lines of code is ~0.15s,
1500 is 0.31s, which is the current size of an average coil program - assuming we recompile the prelude each time.

Some worse case situations (15k recompile) is 6 seconds.

The results vary but we're looking between 2500-3000 LOC/S so far. This is very good especially since the startup time of deno/js is negligible.

And if we start caching compiles for files that don't need re-compiling, a hot-watching compiler would be effectively instant.