we can now compile the prelude!
We now have some more real numbers

the new coil-based compiler takes 0.77s for the prelude, while the clojure based compiler takes 9.68s!!

That's a 12.5x speed improvement!

The compiler still has some bugs, so can't compile itself, but once we get there, it'll be nice & snappy to work in this language again.

While its not necessarily good practice to put all your things in one file, the new compiler, lexer, parser & code gen is just shy of 1400 lines. Which isn't bad!

That is essentially on par with the clojure compiler which is incredible given how concise clojure is and how much of a performance benefit we got from the rewrite