unbounded ranges
we don't have exclusive unbounded  ranges because it overlaps with the spread operator.

This is because the spread operator syntax is very loose to allow it for all the uses javascript allows for.

We could clean this up as there aren't too many cases, I believe its just for function arguments its not being implemented properly.

Maybe even just for function call parsing, since it currently is a simple ~Until[:close_p parse_expr :args].

So yea we could clean that up.. or we could take a note from rust and use .. for exclusive, and ..= for inclusive.

Honestly its very tempting to go for the rust syntax approach because I think .. vs ... is very confusing. I have to remember a saying my old coworker would say is "the last dot pushed out the last element". This is coming from the ruby tradition.

I think this may be a case where we want to borrow from somewhere else.

Its kind of funny how much this language resembles rust given how different the goals are.

On the note of syntax changes.

The object literal syntax is inconsistent with map syntax.

I'm a bit torn on this because with the record syntax doesn't have something similar to {a} which means {a: a}.

Its tough to make an argument for extending the record syntax for this case because while {a: 10} is clear that "a:" means that a is a keyword. Meaning its shorthand for {:a => 10}, {a} is not clear.

But one thing that should be changed for object literal syntax is {[key_expr]: value} instead of {key_expr => value}.

I like that currently object literals are very familiar to current js people, but the arrow syntax is more general & is consistent with the Record syntax.

Things to think about. I'm not 100% sure what the right path is.

I'm feeling strange about this language, each time I make significant progress in it I can't help this overwhelming feeling that its still not enough.

Maybe that feeling keeps me going to work on it, but its frustrating that I can't just accept some satisfaction for what I've done so far.

It also makes it fairly hard to demo this language because while I think there's value here, people don't seem to want to make the leap unless I prove that it is as is ready to replace _your language/framework_. Its not, it may never be, but why is that necessary for it to be a worthwhile & noteworthy project?

Anyways, though this commit is a half broken commit with a compiler that has horrible error messages, I'm quite proud of what I've made.