start to implement OrderedCollection for Range
its not clear the best approach here. I think going fully lazy for take & is nice idea but currently there isn't a unified theme for lazy collections.

I think its something that contains an "Underscore", but I'm still not sold on that.

I like underscore a lot, but it needs a better name.

Or maybe not, its a confusing name, but any other name I can come up with is even more confusing.

At least its easy to look up & has little meaning, so if it gets used in ways different than say a "Transform" would be, idk.

Its not clear, but "Transform" is far too general of a name.

To describe it as clear as I can, it is a data structure which stores functions & arguments. These functions & arguments get applied when you invoke the "Underscore". This makes it feel like a function definition in a way.

Like a (pure, statement-less) function is a series of functions & arguments stored together, and when you call the function it reduces the series of function, argument pairings.

_ is implemented in coil functions which are composed of js functions which are composed of other js functions & c++ native operations which are composed of assembly instructions.

This is one of those things, even though I understand how its implemented, it still confuses me at a fundamental level.

There's always some mystery, I wish more developers would admit that.

I constantly question my design choices, and honestly its only gotten more confusing the more I learn.

I think that's a sign I'm growing.

Less sure about any specific set of design goals.

There are things that I find good that stick around for a while maybe many years but honestly its all up for debate. All of it.

I think of languages like go & elm which reject language-level abstraction. And in real world seem to be one of the better systems for building real, complex ideas.

On the other hand, scala, haskell, rust, even c++ seem to tend to extremely overly complex systems.

I can think of a few reasons why this might to be true

1 - solving program design at the language level is fundamentally impossible.

2 - expressive languages allow for personal freedom which is often let people get carried away.

3 - these languages have overly invasive type systems which encourage pre-maturely over-architected solutions due to the need to fullfil all type errors. I heard a good story from eric normand I believe it was who said described the fundamental difference between a clojure developer & a haskell developer is that clojure developers will give you an opt-out "unsafe" api for unfinished features, where haskell will just give you an incomplete yet type safe API. The example he was talking about was a CURL wrapper lib, and the haskell library didn't have the types for a certain option so it wasn't implemented. This is of course not a fundamental limitation of the language, they have notions of maps & other loosely typed collections. The problem is that it is unidiomatic to provide unsafe/loosely typed arguments.

The other issue with expressive type systems is they tend to be really hard to please, and when you do please them the type errors are cryptic and there is no such thing as a "type-level" breakpoint/debugger.

Where as a counter example, say I'm using a highly dynamic & expressive API in javascript. I can see what line it broke, get a stack trace & have the option to add a "debugger" statement. On top of that, the amount of information available at runtime lets API developers generate quite good error messages (I remember react being surprisingly pleasant). Now this is not always the case, but I think the option to display & manipulate the error messages for you API is much better than having a cryptic type error.

These are all just speaking of the current limitations.

These problems could get solved, but still makes me wonder.

For now I still think coil is worthwhile & it really _feels_ like a language, which makes me feel quite good.

There's a lot to do before its ready for use though and that really bothers me, but I'm not at all excited to work on compiler error messages right now so I won't give it much thought at all. I'll continue doing as I've been doing & just work on what ever is the most interesting at the given moment.

There is no need to _finish_ anything, because that is impossible, there is no target audience for this project, there is no requirements for what determines this to be fruitful other than if it is fun for me to work on. And it is