implemented import & export statements. Very naive & buggy but it works
I'm starting to see benefit in adding a ~ChompId["from"] or something because I'm adding all these reserved keywords and its conflicting with native js apis.

I'm not sure which keywords I want to reserve & which I don't if I take this approach but at some point it'll have to be done because it's getting quite complex.