experiments with idea of a "Repo"
This has A LOT of work ahead of itself, but it would provide something like an ORM like experience interfacing to rest apis.

It should do things like data mapping, handling dependencies across repos & making sure everything is in sync

in the future possibly easy methods to determine if a resource is out of sync with backend.