Clean up prelude.coil a little bit
The prelude is starting to become a pretty big file, and since I recompile it every time, its impacting compile speeds.

Its not a big deal & this is optimized out of the svelte example.

But long term, I want to split up prelude.coil into different files maybe?? I'm still not sure, but it would make it easier for people to find what they are looking for in here, and make incremental compiles faster.

But first I have to make an incremental compiler.. and know how to compile many files in a directory

I have neither of those, so splitting up the prelude right now would only hurt.

Also the build time is 1.2s for an average test file, that's really nothing.

But if that starts to hit 2-3 seconds I'll have to really consider spending the time to work on incremental builds

the compiler is about ~1s/1000 lines, which isn't.. bad. ~ 1.7minutes for a project that's 100k lines, and 17minutes for 1million line codebase.

That's a full build.

For an *large* coil codebase (50k) it should only take 50 seconds, and incremental builds will always* be under a minute.

I don't think there's any point worrying about optimization.

The other part about this to note is that the compiler is 100% parallelizable right now, so running on multiple cores the performance should be very good.

I think I keep getting in my head comparing it to other lines/s compiler speeds, but its fast enough, and there's been 0 optimization work so far so there's likely a fair amount of low hanging fruit.