remove HashMap from prelude.
When we have a proper HashMap, it needs to be well designed & performant. The current HashMap implementation is neither

There's also been no use cases for non primitive keys so far in coil.

Keeping the prelude free of bloat is necessary as it is starting to get large & impact build times.