auto-iter all Iterable protocol methods
this is safe since iterable::iter() === iterable, its one of the rules for iterators in js.

This also makes iterables a lot less friction to use.

I was experimenting with adding a generic ::collect method on all iterables that would know which type they should collect into.

For example: [1, 2, 3]::map(_ + 1)::collect() // [2,3, 4]

This is like a surprising amount of things in the language inspired by rust, but rust afaik uses static analysis to do this.

Anyways, I couldn't seem to get it to work, so for now we'll have to be explicit about converting to arrays or objects or maps or whatever. I honestly don't hate this, but it would be nice to have it one day if it works within the system.