formalize vec destructuring a little more
while I think this vec comp has some juice to it, being a language level dsl makes it hard to useful things without being too verbose for this syntax being worth it.

What would make this theoretically worth it is if we can do something like

~Array[:age + 10 :for {age_int: :age} :of [{age_int: 10}]]

the best we could do is expand the keyword parsing to allow for something like :+, and do some wild parsing inside Array[Vector], but that's a fair amount of work & seems brittle.

what could be better is accepting a variable amount of expressions before :for allowing you to do the following

~Array[:age _ + 10 :for _ :of [{age: 10}]]
still this reads weirdly but it might be on to something..