vec comps starting to feel better.
can compose all fns up to :for, and the lookup expr in between :for and :of is starting to flesh itself out.

It'll soon need to be a protocol though so it can be extended. Its close to the Call protocol, but really it should be called way deconstruct protocol or something.. not clear yet

Also I ran into a minor issue in the prelude with using _::at(:score), this fails because at checks to see if the object impls Vector or Record first & looks up :at on the first one that matches.. in this case it'll look up :at on Vector, so to keep my Iterable functions simple, I just copied the definitions between Record & Vector for Underscore