added conditional support for vec comp
its still a little unclear how useful this will be, but for now, its a fun experiment. I'll wait until I'm finding a real world experiment of wanting this.

Until then this will be a nice experiment in this directory people can reference & take if they want to use it in their code.