experiment for a "random" protocol
This protocol could in future used for something like faker.js to generate random data.

Also being polymorphic, it lets the user add new notions of "random".

I feel like random is almost the wrong word here, a better protocol name might be DataGenerator, but that also feels weird.

I'm sure there's a good name for this but for now we'll stick with random.