implement persistent immutable data structures (Vec & Dict)
Also a new syntax shorthand to use Vec & Dict -

~[1 2 3] -> ~Vec[1 2 3]

~{name: "marcelle"} -> ~Dict{name: "marcelle"}

One thing I've been thinking about is whether or not we should include "..." spread syntax for record & vector literals.

Technically this should be feasible because everything is iterator based.. but it'll be a full copy. So it completely removes the benefits of persistent immutable collections.

Instead I've implemented + for Vec & Dict.

~[:a] + ~[:b] -> ~[:a]::concat(~[:b])

This is not yet stable, and likely not performant. This is a very naive solution, but it seems to work fairly well so far just from usability.

I'll need to figure out how to do performance testing on this. concatenating a lot of Vecs vs Arrays, and spreading a lot of objects vs merging Dicts?

I'll see what's up, and when I get some numbers & feel good about the algorithm I'll put it in the prelude.