Some initial perf results on Vec
when concatenating plain arrays, 10000 iterations is about 0.2s, Vec is 0.1. But when we get to 10000 Vec is 0.3, where arrays is 35 seconds..

So the gains really explode with large lists.

There is likely an inflection point where array copies are faster for small arrays than concatenating Vec, once I find that point I can optimize concatenating Vecs to do copies when the arrays are below that inflection point.

This is just a guess, so I'll have to get the numbers to see.