into experiment
I kind of like the idea of a generic `into` function, expanding on what clojure has.

This lets you pipe any kind of data structure into any other one provided it defines the Into protocol.

One problem off the bat is its awkward to using in a series of ::map's because its polymorphic based off the arguments, not the `this` argument, so we can't use `_`/`Underscore` unfortunately. There might be a way.. but not sure.

Either way its a fun experiment, and stuff like ~Rename really showcase the power & simplicity of coil.