bring (lazy) split into the prelude
I added a static EMPTY property to the Vector protocol.

I was considering extending this to Records, but I'm not sure there's ever a time you need to split something into a record.. vectors should be good enough.

Remove the ToString protocol, it has no clear purpose. Not all strings are the same, there's strings for printing, and strings for serializing those are not the same.

Its not very clear what the purpose of this.toString is besides very simple casts like numbers & things.. but anyways we can figure out what to do there more later. For now I just inlined the protocol because it was only impl'd for Object