Formatter:
Experimenting with ways to determine node size.

Basically after a certain size of a node it should trigger a size breakpoint.

When an array is big it should have trailing commas.

But its not always so clear, for example

[1  2]::map(_ + 1)::map(_ + 2)

the lhs of the first ::map is small, but since its 2 chained maps it should probably wrap to new line

So what I tried is an aggregate of nodes to determine the size, so for a :bind, it takes in to account the lhs size & the expr size.

All together it starts to paint a more holistic natural way to write these formatters

Formatters is one of those tasks that seems like it should be fairly straightforward but its really not

I'm on a mission to make writing it straightforward though because remembering all the rules ad-hoc for each node is very very messy & hard to understand.

If I can only come up with a sane small set of rules to determine wrapping rules then this will become trivial.