Big collections refactor
Now there's a new protocol Collection to take what is shared between Records and Vector

so we have no more (this[Vector] || this[Record])

Its properly defined now

I also started to work on adding runtime available doc strings, idk if this is super useful but its really cool & kind of fun.

Having this stuff at runtime feels useful, even if maybe a little wasteful.

If it becomes not useful for runtime we can just remove the protocol