iterating on new mstate compiler.
feeling good about this direction but still having trouble composing compilers.

The idea is each unit is a full compiler -lexer+parser+emitter, and each part should be individually runnable - ie only run the lexer, only run the parser, or run the whole thing.

That second part is a non goal for now but eventually it should be true.